Conflicts:
	Lilla spelprojektet/Code/D3D11Handler.cpp
	Lilla spelprojektet/Code/Engine.cpp

+++ Lilla spelprojektet/Code/GUI.cpp
@@ -314,7 +314,7 @@ void GUI::createLevelList()
 	while(count != nrOfLevels)
 	{
 		fin >> attribute;
-		for(int i = 0; i < attribute.size(); i++)
+		for(int i = 0; i < (int)attribute.size(); i++)
 		{
 			this->levelList[count].push_back(attribute[i]);
 		}
@@ -344,12 +344,14 @@ void GUI::changeText(D3DXVECTOR2 pos, BUTTONTYPE type)
 	}
 	wchar_t* level = (wchar_t*)levelList[currentLevel].c_str();
 	textBoxes[0].text = level;
+
+
 }
 
 string GUI::getCurrentLevel()const
 {
 	string temp;
-	for(int i = 0; i < levelList[currentLevel].size(); i++)
+	for(int i = 0; i < (int)levelList[currentLevel].size(); i++)
 	{
 		temp.push_back(levelList[currentLevel][i]);
 	}
